+++ src/main/java/com/enjin/enjincoin/sdk/model/service/identities/GetIdentities.java
@@ -2,9 +2,11 @@ package com.enjin.enjincoin.sdk.model.service.identities;
 
 import com.enjin.enjincoin.sdk.graphql.GraphQLRequest;
 
+import java.math.BigInteger;
+
 public class GetIdentities extends GraphQLRequest.Builder {
 
-    public GetIdentities withIdentityId(int identityId) {
+    public GetIdentities withIdentityId(BigInteger identityId) {
         withParameter("id", identityId);
         return this;
     }
