+++ src/main/java/com/enjin/enjincoin/sdk/model/service/identities/UpdateIdentity.java
@@ -2,12 +2,13 @@ package com.enjin.enjincoin.sdk.model.service.identities;
 
 import com.enjin.enjincoin.sdk.graphql.GraphQLRequest;
 
+import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.List;
 
 public class UpdateIdentity extends GraphQLRequest.Builder {
 
-    public UpdateIdentity withIdentityId(int identityId) {
+    public UpdateIdentity withIdentityId(BigInteger identityId) {
         withParameter("id", identityId);
         return this;
     }
@@ -17,7 +18,7 @@ public class UpdateIdentity extends GraphQLRequest.Builder {
         return this;
     }
 
-    public UpdateIdentity withUserId(int userId) {
+    public UpdateIdentity withUserId(BigInteger userId) {
         withParameter("user_id", userId);
         return this;
     }
