+++ src/main/java/com/enjin/enjincoin/sdk/model/service/requests/data/SetApprovalForAllData.java
@@ -3,13 +3,15 @@ package com.enjin.enjincoin.sdk.model.service.requests.data;
 import com.google.gson.annotations.SerializedName;
 import lombok.Builder;
 
+import java.math.BigInteger;
+
 @Builder
 public class SetApprovalForAllData {
 
     private String operator;
 
     @SerializedName("operator_id")
-    private Integer operatorId;
+    private BigInteger operatorId;
 
     private Boolean approved;
