+++ src/main/java/com/enjin/enjincoin/sdk/model/service/users/GetUsers.java
@@ -2,9 +2,11 @@ package com.enjin.enjincoin.sdk.model.service.users;
 
 import com.enjin.enjincoin.sdk.graphql.GraphQLRequest;
 
+import java.math.BigInteger;
+
 public class GetUsers extends GraphQLRequest.Builder {
 
-    public GetUsers withUserId(int userId) {
+    public GetUsers withUserId(BigInteger userId) {
         withParameter("id", userId);
         return this;
     }
